(RE-3586) Make the pl-build-tools repo available during build time
This commit adds in the pl-build-tools repo to debian/ubuntu/el
platforms. For el, we simply add in a new .repo file with the necessary
informaiton. For debian/ubuntu, we wget the pl-build-tools-release
package, install it, and then run apt-get update to make sure the repo
is available.

This will let us pull in dependencies needed for building compiled
software.