(RE-3662) Build in parallel when possible
This commit takes advantage of new functionality added to vanagon to
allow builds to know the number of processes available for them to build
with. Using this information, we can pass it to make to let us have
multiple jobs running at the same time. This will help us speed up build
time. Yay parallel builds!