(RE-3880) Glue some wire together for windows
This commit adds a lovely ruby scripts that builds the windows installer
for puppet-agent. It pulls all the version information to include in the
msi from json files provided in this repo. The script also builds
cfacter for inclusion, so that we always have the most recent version to
include.