(RE-3884) Enable building of AIO on SLES targets
This change set includes all required changes to have SLES 10, 11 and 12
build for i386 and x86_64. The primary changes are the build_requires
package names for cfacter (since the build tools are vanagon generated
for SLES) and updating of the sles platform definitions to include the
build tools repositories.

rubygem-ffi also adds the build_requires of pl-make on SLES 10, as that
make is too old to work with rubygems in this case.