(PUP-2037) Create default environment in packaging
Without this commit, a user will need to do work before a master can be
started using the puppet-agent. The user would need to create the
default directory environment, otherwise puppet will be unable to
compile catalogs for a node. This commit creates the default directory
environment in packaging so it will be present for the server to use.