(maint) Update windows build script
This commit makes two changes.

First, we have updated the cfacter powershell script to take arguments
for the cfacter fork and sha to build. This means that we aren't always
building HEAD of origin/master of the puppetlabs fork when we run the
pwoershell script. This is instrumental in testing, and in ensuring that
we're getting the cfacter build that we want.

Second, this is changing what and how we archive the built cfacter bits.
We were running into an issue where we were having trouble getting
everything we needed for cfacter, including `cfacter.rb`. This commit
moves everything we need to archive into a seperate `archive` directory,
and then archives that. This ensures that we know exactly what is
getting into the archive. Formatting is simpler and creating the archive
is simpler.