(maint) remove cached artifacts in repo
An errant promotion of Puppet from CI added the path directory in the
root of this repo. The additional cruft contains the git checkouts from
bundler of the packaging and vanagon repositories.

Having these repositories in the source tree in this location seems to
interfering with bundler's normal behavior.

This commit removes the files that were accidentally added to the
repository (everything under path/) in commit 6f350b4da while retaining
the update that was the only intended part of that commit (an update to
the puppet.json file).