(RE-4050) Add libblkid-dev for Linux CFacter builds
CFacter has an optional dependency on libblkid, required to enable
several filesystem facts. Without this change, libblkid is not found and
facts will be disabled.

Adds a build requirement in the development version of libblkid to get
headers so CFacter can build against it. libblkid is part of the
util-linux package, so we don't need to ship anything extra.

The build dependency is disabled for systems where we can't get a
development version of the library for headers, such as el-4 and el-5.