(maint) Use apt_repo option to setup Puppet Labs repo
The debian and ubuntu platforms used wget to retrieve repo configuration.
Without this change, the machines may not have wget installed; this is
most likely to occur with Vanagon's docker adapter.

Use the apt_repo setting to specify the repo instead of manually
downloading and configuring it. This removes the implicit dependency on
wget.