(maint) Remove puppet logrotate from puppet-agent
The puppet agent will never log to /var/log/puppetlabs/puppet.log. It
logs by default to syslog, and if it were to be configured to log
somewhere else by the user, we have no idea what that destination would
be. We don't need to logrotate a file that will never exist, so this
commit removes puppet's logrotate rule.