(maint) Relax perms on logdir in puppet-agent
The perms on logdir were inherited from puppet, but they don't need to
be this restrictive. Any components that may be logging sensitive
information can restrict the permissions of their logfiles as needed.
Currently, the puppet agent is not logging to this directory and
mcollective does not log any sensitive information to its logfiles. This
commit relaxes the perms of the logdir to the default of 755.
Additionally, with permissions of 750, other packages such as
puppetserver or puppetdb can't make subdirectories that will be
accessible for their service user (puppet/puppetdb). With permissions of
755, that is not an issue.