(QENG-2001) Use marionette-collective not mcollective
Previously we referred to mco as 'mcollective' in puppet-agent which
was its former package name, while we referred to the project as
'marionette-collective' in the pipeline which is the repository name
on github.

This commit keeps references to 'mcollective' only when referring to
the previous packaging and refers to the project as its github name
elsewhere.