(RE-4039) Update windows specific refs to use 'refs/tags'
Origin works if referring to a branch, but not to a tag. Using
'origin/4.0.0-rc1' will not work, while 'refs/tags/4.0.0-rc1' will. This
commit updates the windows refs to use refs/tags instead of origin for
tags.
The offending commit was
https://github.com/puppetlabs/puppet-agent/commit/3996a00db588bf883905d099a683c5ee8e861cf2,
which changed the previous branch name to a tag without also changing
origin to refs/tags.