(RE-4248) Explicitly name mcollective service
In vanagon, services default to the name of the component. When
mcollective's component name was changed to marionette-collective, this
had the unintended consequence of also changing the service name, which
breaks assumptions in the default file and init script. This commit
explicitly sets the mcollective service name to mcollective to avoid
this problem.