(RE-4168) Pin puppet-agent to a vanagon release
While vanagon development was required for puppet-agent work, it made
sense to pin loosely for the vanagon dependency. Now that vanagon has
support for everything puppet-agent needs at the moment, it makes sense
to pin more conservatively. This commit pins vanagon to a specific git
tag to avoid introducing unintended changes in the packaging.