(FACT-956) Update build to track facter/master
Previously the build was pinned to facter/2.x. We plan to release the
next Puppet-Agent build with native facter, found at facter/master. The
first step is replacing facter/2.x with facter/master on the platforms
where we've enabled building it.

Update the facter component to build native facter and track
facter/master. We preserve a Ruby facter configuration based on
facter/2.x for platforms where we don't yet build native facter.