(RE-4258) Update Gemfile to facilitate testing
In order to test builds of vanagon, we'll be using puppet-agent. Puppet
agent previously had a hard coded version for the vanagon version to use
in the Gemfile. This doesn't work well for testing, so this commit
updates it to use a method borrowed from puppet and modified slightly.
The method allows a jenkins job to override the vanagon version to use
to allow it to test different versions of vanagon.