(RE-4677) Build using the pl-build-tools-vanagon repo
Previously we had branching in the facter component for build
requirements based on which version of linux and which distro you were
using. This removes all of that branching because we have built out the
complete C++ toolchain (gcc, cmake, yaml-cpp, boost) for all of our
supported linux distributions. This means that now we should be using
those for building so that if one of them gets updated, all puppet-agent
platforms will pick up the changes.