(RE-4603,RE-4680) bump augeas to 1.4.0
This version bump includes two lense fixes that we want to have
available. The first is a fix to the mco lense that updates where augeas
looks for client.cfg and server.cfg. The second is an update to the
nginx lense to expand the scope of what the lense can do (i.e. it's now
recursive, can handle nested blocks, etc.)

Rather than patching augeas 1.3.0 to include these fixes, we chose
simply to bump to the latest version (1.4.0), which includes these
updates.