(RE-4545) puppet-agent should use tmpfiles.d w/ systemd
On systemd platforms, the /var/run/puppetlabs directory will not be
automatically recreated after a reboot. We should use the tmpfiles.d
facility to create the /var/run/puppetlabs directory automatically
on reboot.

The way we're doing it now is a total, complete hack where we piggy-
back a tmpfiles.d config file into the Puppet component and use
that to get pieces in the right place. Ideally, this sort of
functionality can be rolled directly into Vanagon itself.