(RE-4826) Apply patch for augeas readline functions
Previously osx would not build augeas 1.4.0 due to missing functions
from libedit (which is used instead of readline on osx). This commit
addresses that by applying a patch to define those functions if they are
not available. The patch is safe to apply everywhere. The patch is based
on https://github.com/hercules-team/augeas/pull/257.