(RE-4879) Generate x86_64 code for 64 bit systems
Previously, openssl unconditionally passed 386 to the configure, which
attempts to create i386 compatible code (bad things happened otherwise
on 486+ processors). This flag is not desireable on 64-bit builds. It
seems to successfully build, but then causes strange problems, for
example in the pkcs12 routines. This commit updates the configure to use
linux-x86_64 for 64-bit builds, and to only pass 386 to the configure to
x86 builds.