(maint) Compile facter with the correct java_home
Prior to this commit, we were pointing 32bit debian and ubuntu builds to
the java-home that corresponded to 64bit builds. Because of this, facter
could not find java, and was not building facter with jruby support,
which means that it wouldn't work with puppetserver. This commit fixes
that by pointing to the correct architecture where java will be found.