(maint) Disable install doc
Prior to this commit we were disabling the documentation
generation/installation for the rdoc components, but not for the C API.
We should probably have all docs, or none. Since this ruby is really
designed to be a bundled ruby with Puppet and not general purpose, I
think it makes the most sense to just disable all documentation.