(RE-5106) Fix osx openssl compilation failures
A recent change to vanagon exposed some failures in the compilation of
openssl on osx related to makedepend. We had been patching it out during
the build, which didn't work when `make depend` ran. This commit removes
that patch, and adds makedepend as a build dependency on osx. It also
adds /usr/local/bin to PATH to ensure that makedepend can be found.

Conflicts:
	configs/components/openssl.rb