(RE-5088) Add custom OSX installer plugin
This commit adds a custom OSX GUI installer plugin that prompts the user
for puppet master server and client cert name.  This is the same plugin
present in the PE agent install.  It also adds additional bits to the
postinstall script that will parse and add the capture input to puppet.conf.