(RE-5290) Update PL ruby components for cross-compile
This commit updates the puppetlabs ruby components for a cross-compile.
Here we use install.rb, but the agent's ruby hasn't been built for the
current system. Because of this, we use cswruby on solaris and pass in
bindir to the install.rb calls. Puppet's install.rb call has been
updated to include --no-check-prereqs to ensure that facter and hiera
are not needed in the csw ruby for install.rb to proceed.