(RE-5290) Update compiled gems for a cross-compile
This commit updates the compiled gems to install against a
cross-compiled ruby. This is a bit odd. Here we use cswruby to load our
cross-compiled ruby's rbconfig to run extconf and generate a makefile
that will use our cross-compiler to build the gem. We also need to unset
CFLAGS because it overrides ruby's CFLAGS, which we want to keep because
it knows better than us.