(maint) Limit chown in ruby-augeas to solaris platforms
Previously a chown was applied to all platforms during the install of
the ruby-augeas component. This was to satisfy solaris packaging needs,
and seemed safe everywhere. However the root group doesn't exist on
platforms such as OSX, so this commit limits the application of the
chown to just solaris where it is needed.