(RE-5290) Add runtime component to puppet-agent
This runtime component is used to copy in libssp, libgcc_s and libstdc++
into our builds for use in the agent. Currently this is limited to
solaris, where we use shared libraries for linking against gcc. This
will be useful as we move more of the stack off of static linking.