(RE-5290) Update ruby for a cross-compile on solaris 10
This commit brings in several updates for ruby to cross-compile
successfully on solaris 10 sparc. First, we add a source of a real
rbconfig from a sparc solaris 10 build. This is swapped in for the
generated rbconfig in the build so that our users can install compiled
gems.
Other updates include setting --with-baseruby and bringing in ruby19 to
use in the cross-compile, adding --host, and disabling rdoc (doesn't get
along with cross compiling). CC is also exported to ensure that the
correct compiler is used.
At the end we also stage the generated rbconfig in the docdir for use
with our compiled gems. A sed call is used to disable a safety check to
prevent us from doing exactly what we're doing.