(RE-5290) Update ruby to replace rbconfig for all solaris
Because we now build solaris all the way through with our gcc, we need
to replace the rbconfig.rb in the package with an rbconfig from a build
against the system gcc. This is because end users won't have our gcc to
build against.
This commit extends the work that was used previously for sparc to
include x86 solaris and adds an rbconfig from a build of ruby 2.1.6 on
solaris 10 x86.