(maint) Fix Solaris i386 build's FACTER_RUBY
Facter relies on FACTER_RUBY being defined to match the AIO's libruby
path. For cross-compiles - such as sparc - we need another Ruby
installed to provide that info, but that Ruby isn't present on Solaris
i386 builds and results in setting the wrong FACTER_RUBY.

Only use `/opt/csw/bin/ruby` on Sparc, as that's the only system it's
needed on.