(RE-5460, RE-5449) Use the runtime component for libgcc
Since we were getting segfaults with facter unalbe to find libgcc and
libstdc++, we need to fix that. The runtime component in this build is a
component that copies over the required libraries from the pl-gcc
package into /opt/puppetlabs/puppet/lib.

We're also add a BR on runtime for augeas, openssl, ruby and facter.