Provide a sane rbconfig for AIX
Since we build ruby using our own GCC and do not ship that gcc, we need
to modify the rbconfig to look for a system compiler. In this case we
default to using the system GCC at 4.2.0 from the AIX Linux toolbox.

The rbconfig is generated using a build of ruby (and thus libedit,
openssl) using stock gcc.