(RE-5401) Add host_ruby setting to puppet-agent
To prevent a lot of duplicated logic, this commit adds a setting called
host_ruby to the puppet-agent project config. This setting indicates the
ruby to use when installing compiled gems, hiera, mco, and puppet, and
when cross-compiling ruby itself.
All such components are also updated to use this new setting.