(RE-5562) Performance improvements for compile times on AIX
Previously a puppet-agent build on AIX 5.3 took about 600 minutes. With
this patch, it takes about 35. All this does is explicitly call
configure with bash for ruby and libedit. Otherwise AIX uses the default
/bin/ksh which somehow results in hundreds-of-thousands of calls to
mktemp and io thrash. I imagine there is something to do with mktemp
built-in the shell vs not, but I haven't looked into it enough to know
the root cause. I do know that 35 minutes vs 600 minutes is a trade we
should make.