(RE-5379) Swap ca-cert for puppet-ca-bundle
This replaces a complete copy of the curl.haxx.se ca-cert bundle with
a stripped down bundle generated from the puppet-ca-bundle repo. This
allows HTTPS access to Amazon Web Services, Microsoft Azure, Puppet
Labs domains, and Rubygems.org while reducing the footprint of the
CA Certs we need to track.

It also removes the conditional logic involved in using the CA Cert
bundle on some platforms, but not others, or linking in local certs
for use by Puppet-Agent's OpenSSL library.