(RE-5358) Update Ruby to 2.1.7 again, but better
As reported in https://bugs.ruby-lang.org/issues/11457, the change
made to how pthread is called in 2.1.7 causes segfaults on some
platforms (we saw it on EL4 and SLES10). This patch corrects fixes
the segfaults by actually ensuring that memory is allocated. This
is backported from upstream Ruby changeset r51626.

Additionally, the static rbconfig files we're maintaining have been
updated and a note has been added about their purpose and their
shortcomings.