(PE-12273) Fix service state on upgrades
Currently, puppet-agent packages prior to 1.2.4 have a bug
in the %preun scriptlet that causes the service to be
unconditionally stopped and unregistered on package upgrade.
As %preun runs *after* the %post of the new package where we
register the service, that leaves us in a broken state.

This adds logic to new packages to:
1. If this is an upgrade, save the state of the services early in the package operation.
2. At the end of all other package operations, restore the state to what it was.