(ARISTA-43) Update puppet config to set codedir
Previously --codedir was not passed to install.rb, which means that
install.rb would default to /etc/puppetlabs/code, which is fine until
arista wants codedir to be /persist/sys/etc/puppetlabs/code, and
/etc/puppetlabs a symlink to /persist/sys/etc/puppetlabs. In that case
not passing --codedir meant that on arista /etc/puppetlabs would still
be created, which makes the symlinking break. This commit adds --codedir
to the install.rb call to ensure that the sysconfdir setting is
respected.