(PE-12686) Remove dependency on tar rpm for AIX
Currently the puppet-agent package depends on having a 'tar'
RPM installed. The tar command is used by PMT, however PNT doesn't
currently work on AIX due to the lack of CA certs. It has not really been
tested, and probably needs some updates to work with AIX.

This removes the useless dependency on the tar RPM package. We
may want to revisit when we update PMT support.