(PA-25) Export RUBYLIB for pure ruby gem installs
Previously, pure ruby gem installs would explode on the cross compiled
ruby because puppet and hiera weren't in the load path, which triggered
a neverending attempt to find them in the gem and ruby load paths,
eventually overflowing the stack.
This commit addresses that problem by explicitly setting RUBYLIB to
include our vendor_ruby directory, which has puppet and hiera. This
allows hiera/version and puppet/version to be safely required and avoids
the aforementioned stack overflow. This will work as long as our version
libraries don't attempt to load any native code, such as facter.