(maint) Dynamically fetch the ruby-stop version
Prior to this commit, if we ever wanted to update the version of ruby
stop, we would have to remember to change the version number in multiple
locations. I don't feel like doing that, since I'm lazy. This commit
update ruby stomp to be able to use the version number we've already
set.