(RE-5734) Include windows specific gems
This adds in a bunch of gems that are specific to the windows platform.
It also adds in the PATH environment variable. Since we're installing
ruby to a different location then the rest of our dlls, the gem install
fails. We need to ensure the gem install can find the rest of the bits
we've installed into our bindir so the gem install succeeds