(maint) Bump MinGW to 5.2.0
In order to switch over to using the nuget packages built with
chocolatey, we need to switch to MinGW 5.2.0. This is mainly due to the
fact the new nuget packages are built with MinGW 5.2.0. It generally
isn't that terrible, but boost uses the GCC version used to compile
libraries in the library name. For CMake to find the boost libraries, we
need to ensure CMake uses the same version of GCC/MinGW to build our
compiled projects that depend on boost as were used to build boost
itself.