(maint) Build the puppet-agent msi with nuget packages
This commit updates the scripts we use to build the puppet-agent MSI. It
switches out the sourses we're using for Boost and Yaml-CPP. Rather then
using precompiled tarballs from our s3 bucket, we now have nupkgs
available in our internal chocolatey feed.

We also set up an additional internal chocolatey feed for packages that
aren't automatically built and shipped. We've populated that with the
rest of our dependencies. As a result, we got to switch over from
installing packages from myget (and other external locations) entirely.

This commit also updates the cmake calls to use the toolchain file we
now have available for cmake. This toolchain file points to where we are
installing a lot of our build time dependencies so cmake can easily find
those libraries.