(PA-138) Build native software on windows with secure components
Prior to this commit, we were building facter, pxp-agent, and
cpp-pcp-client with an outdated version of openssl and with a ruby that
was linked against the same outdated version of openssl.

Given that we are building everything else with a much more recent
version of openssl, we really really should continue to use that.

This commit also updates the paths that facter searches when linking.
We need to ensure all our dependency directories are included in the
search path, including openssl and ruby. This ensures facter will find
and correctly link against those libraries.

Conflicts:
	bin/build-cpp-pcp-client.ps1
	bin/build-facter.ps1
	bin/windows-toolset.ps1