(maint) Switch to cygwin make by default
In our makefile, we occasionally call out to shell to parse some logic.
When we run our makefile with mingw make, these calls just spin. This
causes the build to time out. Which is really sad. Because the majority
of our projects are already using cygwin make, I decided it would be
better if we defaulted to cygwin make, and only used mingw make where we
need to.