(maint) Standardize CMAKE_PREFIX_PATH
This commit standardizes the components we're passing in to the
different projects for cmake prefix path. We have a few default
components, namely curl and openssl, that aren't necessarily required by
all projects. However, it's generally more simple to specify them all
regardless of if they're required. Although this is redundant, it
reduces confusion.