(PA-138) Bump OpenSSL from 1.0.0s to 1.0.2e
This commit bumps OpenSSL from 1.0.0s to 1.0.2e. We need to switch to a
newer version of OpenSSL in order to handle recent security events
(CVE-2015-7551). In order to do this, we need to make sure all OpenSSL
libraries have been updated. Ruby also makes certain OpenSSL libraries
available. The previous Ruby build we were using was making OpenSSL
1.0.0s libraries available. In order to avoid version conflicts, the
switch to OpenSSL 1.0.2e also required us to switch to a different ruby
source. Both the ruby and openssl builds in this commit are the same
that are used in the puppet-agent windows build. The openssl referenced
here is what we are using to build the ruby we make available in
puppet-agent. Likewise, the ruby referenced here is the same ruby we've
built and made available in puppet-agent.