(maint) Update PATH for building
Prior to this commit, PATH was being set incorrectly. This pipe was
joining the different elements we wanted to add to PATH with a space
rather then a semicolon. This commit updates the command to handle that,
so we can correctly set path and access all the elements we've added to
it.