(maint) Remove instances of drive_root
This was a method we were originally using to help us out of the rut of
using /cygdrive/c everywhere as our prefix. I've since fixed some bugs
to where we no longer have to deal with that in most places. As a
result, we've decided to remove the use of drive_root.