(maint) Remove unnecessary branching
Prior to this, we weren't able to specify path beginning with C: for our
windows projects. We have since fixed the bugs preventing that. This now
allows us to remove a lot of conditionals and special casing to handle
if we were building on a windows machine or not. Yay!