(maint) Remove unnecessary branching
This commit removes unnecessary conditionals in puppet-ca-bundle.
Previously, these were required because on windows we had to convert
between /cygdrive/c and C:. I've since fixed a few bugs in vanagon, and
we can now default to the C: prefix. This allows us to remove a lot of
special casing in our build configs.