(maint) Use custom OpenSSL for windows
Unfortunately, the OpenSSL maintainers pack their source code up with
symlinks. Normally, this isn't an issue. It is for windows. MS-DOS
doesn't really have the ability to handle symlinks. Cygwin does, but not
in a way MS-DOS can understand. This leaves us in an interesting
situation. As a result, to unpack and build OpenSSL successfully, we
need to start from a tarball that has had all symlinks removed manually.

Oh joy.