(maint) Remove unnecessary branching
Prior to this commit, we had to do all sorts of special casing and
conditionals to handle the switch between /cygdrive/c and C: as the
prefix for windows paths. We couldn't default to C: due to bugs in
vanagon, and not all of the tools we use understand the /cygdrive/c
format. I've since fixed the bugs preventing us from defaulting to C:.
Now that we're using that prefix, we can remove a lot of special casing
and branching that was previously required for windows