(maint) Update facter directories to be consistent
We should avoid hard coding paths wherever possible. The commit sets the
facts.d dir to be prefixed with a variable we already have set in the
project file. This will allow us to be more consistent with any pathing
updates.

This change also lets us remove a lot of unnecessary logic