(maint) Reduce branching in paths
This commit further reduces the differences in our *nix and windows
paths to the best of our abilities. We want to be fully aware of where
these paths diverge. This commit helps us isolate that.

This commit also updates some of the windows paths to be more in
agreement with what's happening on the *nix platforms.