(maint) Add a rake task to verify tagged components
This commit adds a `release:tag_check` rake task to verify components
are tagged correctly. You'll want to pull the latest puppet-agent before
running the task:

    $ git pull upstream stable
    $ bundle exec rake release:tag_check

If a component refers to a non-existent tag, an error is reported:

    hiera 3.0.7:                                                      [ FAIL ]
      Tag doesn't exist

If a component's tag is not in the standard `refs/tags/1.2.3` form, an
error is reported:

    puppet 4.3.2:                                                     [ FAIL ]
      Tag '4.3.2' is not of the form 'refs/tags/4.3.2'

If a component is not pinned to a tag, an error is reported:

    pxp-agent stable:                                                 [ FAIL ]
      Tag 'stable' is invalid

If a component's tag points to a commit and the commit is not the head
of the stable branch, a warning is reported, along with the SHAs of the
tag and branch:

    cpp-pcp-client 1.0.1:                                             [ WARN ]
      Additional commits on branch 'stable'
        tag    : 9c4db665c5 (1.0.1)
          └->  : 0a90ea7849
        branch : bee31d1b81 (stable)

If a component is missing a URL, an error is reported:

    nssm:                                                             [ FAIL ]
      URL is missing

If a component has a URL, but no `ref`, such as nssm, then the component
is skipped:

    nssm <none>:                                                      [ SKIP ]

Otherwise, if a component's tag points to the head of the branch, then
OK is printed:

    facter 3.1.4:                                                     [ OK  ]

Note some repos don't follow the stable branching scheme, e.g. mco uses
2.8.x, so the rake task compares against the equivalent "stable" branch
for those repos.

The puppet-win32-ruby repo is special since it is effectively two
components, for x86 and x64 ruby, and the rake task will check both.

Github will rate limit unauthenticated requests, so you may need to
create a personal github access token, and set the `GITHUB_AUTH_TOKEN`
environment variable to the value of the token.