(maint) Add elevate.exe to ruby on windows
This is an executable compiled by Josh Cooper and used to run our
projects interactively

We need elevate.exe.config to allow elevate.exe to run against .NET 4.
See
https://github.com/puppetlabs/puppet-win32-ruby/commit/70847bdc0810bc475b5bf31a7485ee47fad8c088
and PUP-1951 for more information.

elevate.exe is used to run our project with elevated permissions. This
doesn't work in a cygwin environment, but otherwise allows facter and
puppet to request UAC elevation and run in a command prompt window with
administrator privileges.