(maint) Update conflicting components after merge-up
Merging master to stable left two components - facter and pxp-agent -
unable to update independently. The old SHAs use a vendored Leatherman,
but a flaw in how they added Leatherman include directories means the
installed Leatherman causes a conflict. That would be fine if it were
just one (that one would pass on update), but since two components have
that issue the one not being updated will cause a build failure.

Update them concurrently to the latest SHA to bypass the issue.