(PA-141) Don't enable pxp-agent service on install
When installing a fresh puppet-agent package on a debian-based system,
our services will be enabled but not running by default. This means that
on restart, our services will start running. This goes against what
happens on rpm systems, which do not start services by default.

This is fine for the most part, but given pxp-agent is a new product and
is not usable outside of PE, we want to ensure it is disabled by default
on a fresh install.

This will not impact users who are upgrading. If they have the pxp-agent
service running, it will restart. If they have the service disabled, it
will remain that way.