(maint) update the service files for windows
This will update the service files and calls to the
install_service method to reflect the fact that vanagon
is now autogenerating values for bin directories and
component group ids. To facilitate this vanagon has added
functionality to grab a service and retrive those values.

The new user experience should be they don't need to know
what values are in ComponentGroup id or the Directory
attribute of the Components, only that they call to
vanagon to grab those values.