(PA-181) Simplify/correct CFLAGS and LDFLAGS logic
The previous version of this change had incorrect LDFLAGS defined for OS X
and not all components safely strip bad values. This has been corected, and
a comment about how LDFLAGS should be set for OS X has been added.

The cascading if/else for setting CFLAGS could lead to more than one situation
no LDFLAGS are set or the standard LDFLAGS are duplicated in each conditional.
We can improve this by setting broad, generally correct defaults and then only
adding to them or changing them as needed. This also eases the pressures that
likely led to a cascading if/else statement.