(RE-6796) Add puppetlabs.ico file to ARP
Add the Puppetlabs Icon to the Add/Remove Programs (control panel).
Also ensure that it and the License.rtf file are placed in the misc
directory on the target installation.

Please note, there is still a remaining task to formally update the
LICENSE.rtf file to the present date (RE-6772)